requirements.txt
Django==5.1
requests==2.32.3
django-allauth==64.0.0
PyJWT==2.9.0
cryptography==43.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
dj-database-url==2.2.0
langchain-community==0.3.0
ollama==0.3.3
faiss-cpu==1.9.0.post1